A few small changes for 16 vs 32 bit, pulling in winsock for network byte ordering and removed some unused local variables.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@5058 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/krb/copy_princ.c
@@ -45,7 +45,7 @@ krb5_copy_principal(context, inprinc, outprinc)
 
     *tempprinc = *inprinc;	/* Copy all of the non-allocated pieces */
 
-    nelems = krb5_princ_size(context, inprinc);
+    nelems = (int) krb5_princ_size(context, inprinc);
     tempprinc->data = malloc(nelems * sizeof(krb5_data));
 
     if (tempprinc->data == 0) {
